refactor: unify condition
